Abstraction!
Updated Entity and InteractEvent to be abstract classes so that they can
have abstract methods (which, basically they already did, it just was
visually unappealing).

Removed unused directives from solution.

Removed Description from IActor and all that inherits it as, for now,
that should either be handled in the LoggingEngine, or by overriding the
ToString() method (so it can just be Monster.ToString() for example).